#7 Move new pag&sort dtos to new directory
